fixed typos and completed refactor
