+++ Jokes/Controllers/JokesController.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
-using Jokes.Data;
+using Jokes.Interfaces;
 using Jokes.Models;
 
 namespace Jokes.Controllers{
@@ -12,13 +12,17 @@ namespace Jokes.Controllers{
         //GET api/jokes
         [HttpGet]
         public ActionResult<IEnumerable<Joke>> GetAllJokes(){
+            var jokeItems = _repository.GetJokes();
 
+            return Ok(jokeItems);
         }
 
         //GET api/jokes/:id
         [HttpGet("{id}")]
         public ActionResult<Joke> GetJokeByID(int id){
+            var jokeItem = _repository.GetJokeById(id); //get id from binding source, uri parameters
 
+            return Ok(jokeItem);
         }
     }
 }
