+++ Jokes/Interfaces/MockJoke.cs
@@ -1,9 +1,9 @@
 using System.Collections.Generic;
 using Jokes.Models;
 
-namespace Jokes.Data{
-    public class MockExample : IJoke{
-        public IEnumerable<Joke> GetAppJokes(){
+namespace Jokes.Interfaces{
+    public class MockJoke : IJokes{
+        public IEnumerable<Joke> GetJokes(){
             var jokes = new List<Joke>{
                 new Joke{Id=0, setUp="What do you call someone without a nose or a body", punchLine="Nobody Knows"},
                 new Joke{Id=1, setUp="What concert costs just 45 cents", punchLine="50 Cent featuring Nickelback"},
@@ -13,8 +13,8 @@ namespace Jokes.Data{
             return jokes;
         }
 
-        Joke GetJokeByID(int id){
-            return new Joke{Id=0, setUp="I don't trust stairs", punchLine="They're always up to something"}
+        public Joke GetJokeById(int id){
+            return new Joke{Id=0, setUp="I don't trust stairs", punchLine="They're always up to something"};
         }
     }
 }
